[ucd/age] Make library no_std
